requirements.txt
openai
langchain
langchain_openai
streamlit
streamlit-chat
tiktoken
python-dotenv
